Moved service to /src
